+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# [2.2.0](https://github.com/chris-yyau/busdriver/compare/v2.1.23...v2.2.0) (2026-09-19)
+
+
+### Features
+
+* **ack-ledger:** opt-in TypeSafe/Jev union classifier for non-review status descriptions ([#870](https://github.com/chris-yyau/busdriver/issues/870)) ([f81b45b](https://github.com/chris-yyau/busdriver/commit/f81b45b59c48cd6491cecb30601edaa5bcb7457c)), closes [#709](https://github.com/chris-yyau/busdriver/issues/709) [#325](https://github.com/chris-yyau/busdriver/issues/325) [#325](https://github.com/chris-yyau/busdriver/issues/325)
+
+## [2.1.23](https://github.com/chris-yyau/busdriver/compare/v2.1.22...v2.1.23) (2026-09-18)
+
+
+### Bug Fixes
+
+* **harness-audit:** fail closed on unparseable suggest-compact shell registrations ([#871](https://github.com/chris-yyau/busdriver/issues/871)) ([b5021ec](https://github.com/chris-yyau/busdriver/commit/b5021ecab4adf5ac02c6999c1a59516f24c26b8b))
+
 ## [2.1.22](https://github.com/chris-yyau/busdriver/compare/v2.1.21...v2.1.22) (2026-09-18)